requirements.txt
pyspark
boto3
psycopg2-binary
pyyaml
pytest
pytest-spark
numpy
setuptools